requirements.txt
cryptography==3.3.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.4.2
pylint==2.5.2
python-dotenv==0.13.0
pytz==2020.1
query-string==2019.4.13
requests==2.23.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
itsdangerous==1.1.0
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
cryptocompare